requirements.txt
bcrypt
pandas
streamlit
sqlite3
Pillow
smtplib
qrcode
